requirements.txt
influxdb-client
pygazpar==1.1.2
paho-mqtt
